Commit before buffer type alteration
Christov struct implemented but not tested. Still getting erros
regarding buffer incompatibility (uint16 vs uint32), will alter that now

+++ ADC_DMA_1/Core/Inc/main.h
@@ -87,10 +87,10 @@ typedef struct {
 
     float F;
 
-    int R;
-    int RR[5];
+    float R;
+    float RR[5];
     int rr_index;
-    int Rm;
+    float Rm;
 
     float MFR;
